% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_comm.R
\name{calc_comm}
\alias{calc_comm}
\title{Calculate commisions on trade execution}
\usage{
calc_comm(strat, close, profit)
}
\arguments{
\item{strat}{string of strategy run}

\item{close}{numeric of the trade close price}

\item{profit}{numeric of the trade profit prior to commissions}
}
\description{
{
calc_comm calculates commission fees for trades based on TD Ameritrades fee
structure
}
}
\examples{



# Function arguments:-----------------------------------------------------------
}
