% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mrgn.R
\name{calc_mrgn}
\alias{calc_mrgn}
\title{Calculate initial margin on opening trades}
\usage{
calc_mrgn(strat = NULL, prc = NULL, spstrk = NULL, lpstrk = NULL,
  scstrk = NULL, lcstrk = NULL, spcred = NULL, lpcred = NULL,
  sccred = NULL, lccred = NULL, qty = 1, shares = 100)
}
\description{
{
calc_mrgn calculates the opening margin required depending on strategy type
}
}
\examples{


}
