% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_html.R
\name{output_HTML}
\alias{output_HTML}
\title{Simple function to format HTML output to pass back to shiny}
\usage{
output_HTML()
}
\description{
{
output_HTML Called by the server.R script to send HTML output to screen
}
}
\examples{

}
