% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_money.R
\name{print_money}
\alias{print_money}
\title{Format numeric to currency for output}
\usage{
print_money(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
numeric with 2 decimal places and big.mark comma
}
\description{
{
print_money formats numerics to currency for out put in shiny app
}
}
\examples{
print_money(14857)
returns 14,857.00

}
