% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_trade.R
\name{tooltip_trade}
\alias{tooltip_trade}
\title{Generating tooltip (hover over) text for plots}
\usage{
tooltip_trade(x)
}
\arguments{
\item{x}{data.frame of trade results}
}
\value{
string of HTML formated values formated from the specific results row
to be included in the hover over of a plot
}
\description{
{
tooltip_trade takes a results data.frame from a study and selects the row that
you are hoveringformats numerics to currency for out put in shiny app
}
}
\examples{
tooltip_trade(results)

}
